requirements.txt
bleach==1.5.0
bluelens-spawning-pool==1.0.9
boto3==1.4.7
botocore==1.7.17
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.0.11
dill==0.2.7.1
docutils==0.14
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-auth==1.1.1
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.15
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.4
grpcio==1.7.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
imageio==2.1.2
jmespath==0.9.3
Markdown==2.6.9
matplotlib==2.0.2
monotonic==1.3
moviepy==0.2.3.2
numpy==1.13.1
oauth2client==3.0.0
olefile==0.44
Pillow==4.2.1
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.4.0
pyasn1==0.3.6
pyasn1-modules==0.1.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
six==1.11.0
stylelens-index==1.0.0
stylelens-search-vector==1.0.2
tenacity==4.5.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.6
tqdm==4.11.2
urllib3==1.22
Werkzeug==0.12.2
